GCD: Move Memory Block Init to Separate Function
Currently, when the PHIT HOB is processed to update the GCD,
it calls add_memory_space() which sees the GCD memory blocks
capacity is 0, so it calls init_memory_blocks, which creates
capacity in the memory blocks and then calls back into
add_memory_space to add the region to the GCD, which this time
sees the capacity is non-zero and so does not call init_memory_blocks
again.

This commit breaks the call to init_memory_blocks out into a
separately callable function that is called when the PHIT HOB is
processed. This is done to simply this call stack, most importantly
because it has been observed that speculative execution can mispredict
the capacity == 0 branch for other add_memory_space calls and do
speculative reads of MMIO (when it believes that this memory region will
be the one that is added to initialize the memory blocks) that on some
platforms can cause stuck transactions (e.g. if some of the MMIO isn't
back by an actual device or memory).

It is the responsibility of the platform to correctly report its memory
resources, but, we can be a bit more defensive here since an active
issue was seen and this pattern was convoluted anyway.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>